separate newline style from whitespace style
New "ui.virtual.newline" style is used for newline characters.
Will fallback on "ui.virtual.whitespace" if not defined.

Useful when newline render is 'after-whitespace', since in this case
you want it to stand out, whereas other whitespace characters would
need to blend in the background.

It was also conflicting with my theme where string literals are
underlined, and the "ui.virtual.whitespace" fg color was applied
on the underline on whitespaces inside strings.

Example theme:
    "ui.virtual.whitespace" = { fg = "gray", modifiers = ["dim"] }
    "ui.virtual.newline" = { fg = "red", modifiers = ["bold"] }